Navigation Visual Style Overhaul
Navigation now has proper support for page highlighting (though it will
need to be updated manually for each page added). The fonts are how
they are supposed to be, and the header has been repurposed to hold an
icon of the project or other highlighted object, and presumably a quote
to go with it. Still needs work though.