Added preliminary mobile support, project image
Uses @media queries to change some spacing. Still needs a lot of work,
but since it is functional at this point, I'm pushing this as a
preliminary release. I'll still need to attach the project pages, but
that will be coming later when I actually have a final build for any of
those projects.