Big style changes, comments, and a new post explaining it all
Removed the background, replaced with more whitesmoke color, made a
proper footer, added comments by Disqus. changed the padding and
margins on a number of objects, and the link styles have changed too.
The paragraph font is not 0.1em larger, and at least the iPhones are
all supported on mobile. Also, new About page to replace contact page.